add dto to get url

+++ models/url.go
@@ -13,7 +13,7 @@ type Tinyurl struct {
 	OriginalUrl string    `bun:"original_url,notnull" json:"originalUrl"`
 	ShortUrl    string    `bun:"short_url,unique,notnull" json:"shortUrl"`
 	Comment     string    `bun:"comment" json:"comment"`
-	UserID      int       `bun:"user_id,rel:belongs-to,join:id" json:"userId"`
+	UserID      int64     `bun:"user_id,rel:belongs-to,join:id" json:"userId"`
 	ExpiredAt   time.Time `bun:"expired_at,notnull" json:"expiredAt"`
 	CreatedAt   time.Time `bun:"created_at,nullzero,notnull,default:current_timestamp" json:"createdAt"`
 	CreatedBy   string    `bun:"created_by,notnull" json:"createdBy"`
